requirements.txt
asgiref==3.9.1
Django==5.2.5
django-simple-history==3.10.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
